+++ apps/meteor/app/api/server/index.ts
@@ -9,6 +9,7 @@ import './helpers/parseJsonQuery';
 import './default/info';
 import './v1/assets';
 import './v1/calendar';
+import './v1/call-history';
 import './v1/channels';
 import './v1/chat';
 import './v1/cloud';