+++ apps/meteor/client/views/room/Header/RoomHeader.tsx
@@ -29,12 +29,12 @@ export type RoomHeaderProps = {
 			pre?: ReactNode;
 			content?: ReactNode;
 			pos?: ReactNode;
+			hidden?: boolean;
 		};
 	};
-	roomToolbox?: JSX.Element;
 };
 
-const RoomHeader = ({ room, topic = '', slots = {}, roomToolbox }: RoomHeaderProps) => {
+const RoomHeader = ({ room, topic = '', slots = {} }: RoomHeaderProps) => {
 	const { t } = useTranslation();
 
 	return (
@@ -64,13 +64,15 @@ const RoomHeader = ({ room, topic = '', slots = {}, roomToolbox }: RoomHeaderPro
 				)}
 			</HeaderContent>
 			{slots?.posContent}
-			<Suspense fallback={null}>
-				<HeaderToolbar aria-label={t('Toolbox_room_actions')}>
-					{slots?.toolbox?.pre}
-					{slots?.toolbox?.content || roomToolbox || <RoomToolbox />}
-					{slots?.toolbox?.pos}
-				</HeaderToolbar>
-			</Suspense>
+			{slots.toolbox?.hidden !== true && (
+				<Suspense fallback={null}>
+					<HeaderToolbar aria-label={t('Toolbox_room_actions')}>
+						{slots?.toolbox?.pre}
+						{slots?.toolbox?.content || <RoomToolbox />}
+						{slots?.toolbox?.pos}
+					</HeaderToolbar>
+				</Suspense>
+			)}
 			{slots?.end}
 		</Header>
 	);