+++ packages/ui-voip/src/index.ts
@@ -3,3 +3,5 @@ export { default as MediaCallProvider } from './context/MediaCallProvider';
 export { MediaCallContext, useMediaCallExternalContext as useMediaCallContext, type PeerInfo } from './context';
 
 export { useMediaCallAction } from './hooks';
+
+export { getHistoryMessagePayload } from './ui-kit/getHistoryMessagePayload';