+++ apps/meteor/ee/server/services/CHANGELOG.md
@@ -1,5 +1,19 @@
 # rocketchat-services
 
+## 2.0.36
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.13.1
+  - @rocket.chat/rest-typings@7.13.1
+  - @rocket.chat/core-services@0.11.1
+  - @rocket.chat/model-typings@1.9.1
+  - @rocket.chat/models@1.8.1
+  - @rocket.chat/network-broker@0.2.24
+  </details>
+
 ## 2.0.35
 
 ### Patch Changes