+++ ee/packages/license/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocket.chat/license
 
+## 1.1.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.13.1
+  </details>
+
 ## 1.1.4
 
 ### Patch Changes