+++ packages/cron/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rocket.chat/cron
 
+## 0.1.45
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.13.1
+  - @rocket.chat/models@1.8.1
+  </details>
+
 ## 0.1.44
 
 ### Patch Changes