+++ packages/ui-contexts/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rocket.chat/ui-contexts
 
+## 25.0.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.13.1
+  - @rocket.chat/rest-typings@7.13.1
+  - @rocket.chat/ddp-client@0.3.45
+  </details>
+
 ## 25.0.0
 
 ### Minor Changes